emacs: Add use-package bootstrap.
This is the beginning of the 4th rewrite of my emacs config.

This version assumes that a full network connection exists (how else
would you clone the git repository?), and that all packages are to be
installed and configured automatically on startup.

The previous philosophy of a perfectly gracefully degrading
configuration was too hard to adhere to; there were too many possible
error scenarios to account for.