emacs: Handle `:general` section aggregation.
As of some recent version of `use-package`, all blocks using a given
keyword (e.g. `:general`) are aggregated into a single list before being passed
to the corresponding keyword handler.

Ostensibly, this is done to reduce the number of calls to the keyword handler,
which is probably does. However, this might mean that multiple calls to the
keyword handler might not behave the same as before, as with `:general`. This
can be verified by expanding the `use-package` macro manually, and comparing the
results.

The solution is to skip using multiple keyword blocks, and pass multiple forms
to the same keyword.