airlift: Pin nixpkgs to a specific revision.
This seemingly simple change came as the result of a number of
frustrating experiments. The only way to achieve a decent level of
reproducibility is to explicitly name a revision of ~nixpkgs~ by
hash. This means that we are guaranteed the same package set until the
hash changes, and that the Dockerfile will rebuild when the hash does
change.

However, this interacts poorly with binary caching, for which I have no
answer yet. Ideally, we should ensure that the nix store used to build
the airlift image is cached across builds, somewhere persistently, and
occasionally collected using the latest airlift package set as roots (or
never?).