[airlift] Remove ansible dependency, add nix conf.
Ansible is proving to be too complex just for deploying a bunch of
dotfiles, given that nix is doing the heavy lifting for installation.

The original premise for ansible was the ability to deploy airlift to
a remote machine without needing to clone the repository on that host.
This is also proving to be less and less useful, seeing as the
deployment depends heavily on software deployed by nix, which is
unlikely to be done without the permissions necessary to clone the
repository on the remote host in the first place.